Fix Circular ESLint Plugin `stringify` Error
Previously, attempting to use an overridden config containing a plugin with a circular structure (as is suggested for flat configs) would error, as the caching implementation would attempt to `stringify` said plugins in order to use them as cache keys.

This commit fixes the issue by replacing the cache object with a `Map`, and using the plugin objects directly as keys, for whom equality can be checked even in the presence of circular references.

A test case (verified to error without the fix) is also included.